refactor: move rate limiter configuration read into its own file. tidy exports in index.ts for rate limiter
refactor: move rate limiter config file to /lib and tidy exports from rate-limiter/index.ts

refactor: move rate limiter state from requestState.ts into the middleware itself

refactor: simplify config file read method
