+++ __mocks__/fileMock.js
@@ -1,1 +1,3 @@
+// This is used to allow our Supertest Jest suite to run successfully (see jest.config.js)
+// eslint-disable-next-line functional/immutable-data
 module.exports = '';
