+++ src/server/index.ts
@@ -9,8 +9,10 @@ const server: Express = express();
 
 applyMiddleware(server);
 
-export default server.listen(port);
+const serverInstance = server.listen(port);
 
 server.set('trust proxy', true);
 
 logger.info(`server running on port ${port}`);
+
+export default serverInstance;
