+++ src/server/lib/rate-limit/index.ts
@@ -1,8 +1,6 @@
 import rateLimit from './rateLimit';
 
-import { RateLimiterConfiguration as _RateLimiterConfiguration } from './validateConfiguration';
-
-export type RateLimiterConfiguration = _RateLimiterConfiguration;
+export type { RateLimiterConfiguration } from './configurationValidator';
 
 export * from './types';
 export default rateLimit;
