+++ src/server/models/Express.ts
@@ -32,12 +32,6 @@ export interface RequestState {
   ophanConfig: OphanConfig;
   sentryConfig: SentryConfig;
   browser: Bowser.Parser.Details;
-  rateLimitData: {
-    ip?: string;
-    email?: string;
-    accessToken?: string;
-    oktaId?: string;
-  };
 }
 
 export interface ResponseWithRequestState extends Response {
